threatlens/requirements.txt
streamlit>=1.34.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.20.0
python-dotenv>=1.0.0
